Mod: Changed min, max and distance property datatype on model and migration files

+++ src/models/combination.model.js
@@ -19,15 +19,15 @@ function default_1(sequelize, dataTypes) {
             allowNull: true
         },
         min: {
-            type: dataTypes.INTEGER,
+            type: dataTypes.DECIMAL(10, 2),
             allowNull: true
         },
         max: {
-            type: dataTypes.INTEGER,
+            type: dataTypes.DECIMAL(10, 2),
             allowNull: true
         },
         distance: {
-            type: dataTypes.INTEGER,
+            type: dataTypes.DECIMAL(10, 2),
             allowNull: true
         },
         status: {
